# frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Sistema de Ponto",
  "description": "\nAPI do sistema de ponto eletrônico com reconhecimento facial.\n\n* **Administrador**: `Authorization: Bearer <token>` obtido em `/auth/login`.\n* **Terminal (kiosk)**: `Authorization: Device <token>` obtido ao cadastrar o dispositivo.\n* Instantes são devolvidos no fuso da empresa; períodos (`date_from`/`date_to`) são inclusivos.\n* Erros seguem o formato `{\"error\": {\"code\", \"message\", \"details\", \"request_id\"}}`.\n",
  "version": "0.3.0"
 },
 "paths": {
...
  "/api/v1/kiosk/identify": {
   "post": {
    "tags": [
     "kiosk"
    ],
    "summary": "Identify",
    "description": "Reconhece o funcionário pelo rosto. A imagem é descartada ao fim da requisição.",
    "operationId": "identify_api_v1_kiosk_identify_post",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_identify_api_v1_kiosk_identify_post"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/IdentifyOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/kiosk/records": {
   "post": {
    "tags": [
     "kiosk"
    ],
    "summary": "Create Record",
    "description": "Registra a batida escolhida. O horário é o do servidor; o token é consumido.",
    "operationId": "create_record_api_v1_kiosk_records_post",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/KioskRecordIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/KioskRecordOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/kiosk/hour-bank": {
   "post": {
    "tags": [
     "kiosk"
    ],
    "summary": "Own Hour Bank",
    "description": "Banco de horas do funcionário reconhecido — e somente dele (BUSINESS-RULES.md §9.1).",
    "operationId": "own_hour_bank_api_v1_kiosk_hour_bank_post",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/KioskHourBankIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/KioskHourBankOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "Body_identify_api_v1_kiosk_identify_post": {
    "properties": {
     "image": {
      "type": "string",
      "contentMediaType": "application/octet-stream",
      "title": "Image",
      "description": "Um quadro da câmera (JPEG/PNG, até 1 MB)"
     }
    },
    "type": "object",
    "required": [
     "image"
    ],
    "title": "Body_identify_api_v1_kiosk_identify_post"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "IdentifyOut": {
    "properties": {
     "identification_token": {
      "type": "string",
      "title": "Identification Token"
     },
     "expires_in": {
      "type": "integer",
      "title": "Expires In"
     },
     "employee_name": {
      "type": "string",
      "title": "Employee Name"
     },
     "workday_date": {
      "type": "string",
      "format": "date",
      "title": "Workday Date"
     },
     "allowed_types": {
      "items": {
       "type": "string",
       "enum": [
        "ENTRY",
        "LUNCH_EXIT",
        "LUNCH_RETURN",
        "EXIT"
       ]
      },
      "type": "array",
      "title": "Allowed Types",
      "description": "Botões de batida que devem ficar habilitados"
     },
     "hour_bank_screen_seconds": {
      "type": "integer",
      "title": "Hour Bank Screen Seconds"
     }
    },
    "type": "object",
    "required": [
     "identification_token",
     "expires_in",
     "employee_name",
     "workday_date",
     "allowed_types",
     "hour_bank_screen_seconds"
    ],
    "title": "IdentifyOut"
   },
   "KioskDay": {
    "properties": {
     "date": {
      "type": "string",
      "format": "date",
      "title": "Date"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "worked_minutes": {
      "type": "integer",
      "title": "Worked Minutes"
     },
     "balance_minutes": {
      "type": "integer",
      "title": "Balance Minutes"
     },
     "counts_for_bank": {
      "type": "boolean",
      "title": "Counts For Bank"
     }
    },
    "type": "object",
    "required": [
     "date",
     "status",
     "worked_minutes",
     "balance_minutes",
     "counts_for_bank"
    ],
    "title": "KioskDay"
   },
   "KioskHourBankIn": {
    "properties": {
     "identification_token": {
      "type": "string",
      "maxLength": 200,
      "minLength": 10,
      "title": "Identification Token"
     }
    },
    "type": "object",
    "required": [
     "identification_token"
    ],
    "title": "KioskHourBankIn"
   },
   "KioskHourBankOut": {
    "properties": {
     "employee_name": {
      "type": "string",
      "title": "Employee Name"
     },
     "balance_minutes": {
      "type": "integer",
      "title": "Balance Minutes",
      "description": "Saldo acumulado do banco de horas até hoje"
     },
     "month_start": {
      "type": "string",
      "format": "date",
      "title": "Month Start"
     },
     "month_overtime_minutes": {
      "type": "integer",
      "title": "Month Overtime Minutes"
     },
     "month_missing_minutes": {
      "type": "integer",
      "title": "Month Missing Minutes"
     },
     "month_absences": {
      "type": "integer",
      "title": "Month Absences"
     },
     "pay_period_start": {
      "type": "string",
      "format": "date",
      "title": "Pay Period Start"
     },
     "pay_period_end": {
      "type": "string",
      "format": "date",
      "title": "Pay Period End"
     },
     "pay_period_overtime_minutes": {
      "type": "integer",
      "title": "Pay Period Overtime Minutes"
     },
     "pay_period_missing_minutes": {
      "type": "integer",
      "title": "Pay Period Missing Minutes"
     },
     "days": {
      "items": {
       "$ref": "#/components/schemas/KioskDay"
      },
      "type": "array",
      "title": "Days"
     },
     "screen_seconds": {
      "type": "integer",
      "title": "Screen Seconds"
     }
    },
    "type": "object",
    "required": [
     "employee_name",
     "balance_minutes",
     "month_start",
     "month_overtime_minutes",
     "month_missing_minutes",
     "month_absences",
     "pay_period_start",
     "pay_period_end",
     "pay_period_overtime_minutes",
     "pay_period_missing_minutes",
     "days",
     "screen_seconds"
    ],
    "title": "KioskHourBankOut"
   },
   "KioskRecordIn": {
    "properties": {
     "identification_token": {
      "type": "string",
      "maxLength": 200,
      "minLength": 10,
      "title": "Identification Token"
     },
     "type": {
      "type": "string",
      "enum": [
       "ENTRY",
       "LUNCH_EXIT",
       "LUNCH_RETURN",
       "EXIT"
      ],
      "title": "Type"
     }
    },
    "type": "object",
    "required": [
     "identification_token",
     "type"
    ],
    "title": "KioskRecordIn"
   },
   "KioskRecordOut": {
    "properties": {
     "employee_name": {
      "type": "string",
      "title": "Employee Name"
     },
     "type": {
      "type": "string",
      "enum": [
       "ENTRY",
       "LUNCH_EXIT",
       "LUNCH_RETURN",
       "EXIT"
      ],
      "title": "Type"
     },
     "recorded_at": {
      "type": "string",
      "title": "Recorded At"
     },
     "workday_date": {
      "type": "string",
      "format": "date",
      "title": "Workday Date"
     }
    },
    "type": "object",
    "required": [
     "employee_name",
     "type",
     "recorded_at",
     "workday_date"
    ],
    "title": "KioskRecordOut"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 },
...
}